Store marks to file so that the ProcessCache system works
Branches and marks are now long-lived.  The marks have to survive even
if the fast-import process goes away.  So, use the --import- and --export-marks
feature of fast-import to persist marks.

These marks are only meant for the duration of this 'svn-all-fast-export' and
don't confer any new incremental behaviour.  You'll need a mark to SVN commit map
for that, at least.