Improve determination of branch point
This fixes the logic of commit 209e6ce4ddf114494d6d72455690af819dcbf18c

qLowerBound doesn't have the desired semantics -- it returns the position of the
first item which is not smaller than the search key.

The intended semantics seems to be to find the given key, or the next smallest one.
This is almost given by qUpperBound -- it returns an element one past the desired result.

I've also added some additional debugging code to help debug this.