Track commits per-branch
Previously, all the SVN commits were tracked in a linear array, and we searched
for nearest commits in that array.  However, SVN history is not linear, and the
'next smallest commit' search was picking the wrong commit.

I've moved the commit array into the 'Branch' structure.

As a minor subtlety, the branch creation revision is also noted in the
'commitMarks' structure, by copying the commit mark of the branch point.
We need this to ensure that the commits array is strictly non-decreasing.