Improve progess log
Make the progress log for commit and branch creation more uniform and
machine parse-able.  Add 'mark' information to log so that it can be
cross-referenced against the fast-import marks file.

The log-* files now have enough information in a convenient enough form
that the export can eventually be made fully resumable and incremental.
The format is essentially

  progress SVN r<svn-rev> branch <git-branch> = (:<mark> | <other-git-branch>)

with a possible trailing # comment.  Any line not matching this can be

The incremental mode would prime the internal structures with

  if (<mark>) repository->commitMarks[<svn-rev>] = <mark>;
  repository->branches[<git-branch>].commits.append(<svn-rev>);