Handle SVN directory deletes that lead to branch deletes
SVN directory deletes often indicate one or more branch deletions.  However,
since the deleted directory isn't present in the resulting revision, several
of the indicators used by the rule-application mechanism aren't present.
This forces us to introduce several useless dummy rules to avoid errors.

We now note deletions and use the previous revision to determine several
properties, including whether the deleted item is a directory or not, and to
enumerate the contents of the directory in recurse mode.

We add an additional heuristic for unknown repositories -- i.e., when a rule
fired, but it's repository was invalid.  We recurse in this case hoping to
catch a more specific rule.  I believe this is safe: because some other rule
must've seen the same directory before, when it or a subdirectory was created,
and decided _not_ to create a repository at that point -- so recursing and/or
ignoring the contents of the just deleted directory won't corrupt the history,
it can only improve it.

We use mark :0 to note mark deletions internally, and in the progress logs.

(Note that cvs2svn creates wierd commits where a whole tree is copied first,
and then subtrees are pruned.  In such cases, neither the previous revision
nor the current revision have the deleted directory -- we ignore this case
as before.  There's no information loss since the final contents of the revision
are exactly what is desired.)