Introduce incremental mode with --incremental flag
We use the progress logs that we carefully maintained to recreate the
branch data structures, as described in earlier commits.

A major change/improvement is that reloadBranches() now uses the marks file
and internal data structures to prime the fast-import rather than using
git-rev-parse.

We also handle --resume-from properly, by truncating the log file to revisions
that only precede the revision resumed from.  Note that git fast-import allows
marks to be reused without any extra processing.