Reduce size of fast-import marks file by not persisting file-level marks
Use two allocators for marks, one persistent commit counter that starts at 0
and counts up commits, and a transitional counter, for files, that counts down
from maxMark, and is reset on each SVN revision.

Note that the marks file will still have marks for some, but not all, files.
The number of such marks is limited by the size of the SVN revision that affects
the most files.

For instance, this changed the size of one marks file from 19M to 3.2M.

fast-import issues: We currently set maxMark = (1<<20)-1.  Anything large seems
to trigger a bug in the sparse array dumping routine in git-fast-import in
certain versions of git.