make --incremental robust to inconsistent import directories
An interrupted import (say with Ctrl-C) can leave the import directory in an
inconsistent state.  This can be due to checkpointing fast-import only
occassionally, but updating log-* files immediately, and/or other reasons.

The incremental mode can detect certain such situations and rewind back to a
safe state.  Note that since the default commit-interval is quite large, this
rewind can end up backtracking a lot.

Note also that import interrupted under the control of svn2git, say, for
missing rules should leave the import directory in a consistent state for
the purpose of svn2git.