Make error handling of --incremental and --resume-from idempotent
When --resume-from failed in incremental mode, the log files that detected
the error condition were truncated.  So, if the same command line was executed
again, the invocation would go through.

We now restore the log files from backup when we detect we're going to fail.
The restored log files may not all be the same as we originally started with,
but we only truncate information that would anyway be truncated on the
next successful run.