Handle branching better for branches that use "prefix" rules
When creating branches/tags off a branch that use "prefix" rules,
svn2git used to create new commits rather than just copying the git
ref over.  This was due to SvnRevision::exportInteral() using
'path.isEmpty()' to determine if a change related to the whole branch
or not.  If a "prefix" rule is used, then 'path' will not be empty.

We change it to instead look at how much of the change string was
chomped up by the rule match.  If the whole string was chomped up, it
means that the change describes the whole branch.