Move backup branches and tags to refs/branches/
The naming scheme is

  refs/backups/r<svn-revision>/(heads|tags)/<branch>

Where <svn-revision> is the revision where the branch is being reset:
either for deletion or to be overwritten.

We use a separate namespace so that we don't clutter up branch-name lists
and tag lists with deleted tags.  These refs will keep the commits alive
as far as 'git gc' is concerned, but will be fairly unobstrusive otherwise.