Provide a way to merge repositories
Suppose you have multiple repositories in SVN that you want to merge into
a single one in GIT, it can get very messy to handle all the special-case
rules.

Instead, we introduce a new "forwarding repository" concept, which looks like

  repository subordinate
	repository unified
	prefix foo/
  end repository

This forwards all commits on the "subordinate" SVN tree to the "unified" GIT
tree, with each file prefixed with "foo/".