Save deleted branches in a visible namespace
 - The branch is first created properly and then deleted
   (line "progress SVN r77 branch config-header = :0 # delete")

 - Jehan's post-1.0.3 commit 584005f2e26149282aa12c4a0367d250caaf3918
   writes a backup of the branches latest ref to refs/backup/
   on deletion.  Many thanks for that commit!

However, in my opinion the problem with the current code is that it
makes it much too easy to lose history during the conversion to Git.
Without a tag or branch pointing to it the commits will fall off easily,
especially if people use "git clone" on the bare Git repository created
by svn2git.

It would be safer and useful to create visible tags for this, e.g.
refs/tags/branch-bugfix17-deleted.

My attached patch changes Jehan's code to do just that.
Please consider application - this patch really matters to me.
If you'd rather have both a ref in refs/backup/ and in refs/tags/ or
require a command line option for it I can adjust the patch for you.

Thanks for listening,