Abort on conflicting ref names
If you have a ref name refs/tags/foo you cannot create a ref refs/tags/foo/bar
and also the other way around, as for the former a file and the the latter a directory
would be created that both would be named 'foo' as long as the refs are not packed.

So check for this and abort early so that the rules can be adapted to produce proper names.