Fix #43: exported path-replace behavior inconsistent with svn
SVN treats a folder replace as a deletion + addition. svn2git was
treating it simply as an addition.

Thus if f2 was supposed to replace f1, the result was the union
of the previous content of f1 and the content of f2.

  r1: A f1/file1.txt
  r2: A f2/file2.txt
  r3: R f1 from f2

'svn log' actually records this as:

  r3: D f1
      A f1  (from f2)

SVN Result:
   f1/file2.txt
   f2/file2.txt

svn-all-fast-export result:
   f1/file1.txt  << WRONG
   f1/file2.txt
   f2/file2.txt